apps/ renamed to utils/, login required decorator added
